Added an exponential backoff to graphics device resetting.
Modified XnaForm and FullScreenTransparentForm to have an AutoInvalidate option in constructor (true by default). If set to false, window will not be invalidated in every update loop. Modified WebViewForm to set this to false and then only invalidate the screen when some rendering is done by awesomium. This will reduce the drawing frame rate to a minimum. Could further improve on this by only invalidating sections of the screen, but this probably wouldn't help much.

Signed-off-by: Martin Evans <martindevans@gmail.com>